+++ player/amiga/mr_master_options.h
@@ -112,6 +112,7 @@ static inline int mr_master_options_apply(mr_play_options *options)
     options->h264_performance = master.h264_performance;
     options->audio_rate = master.audio_rate;
     options->no_audio = master.no_audio;
+    options->mono_audio = master.mono_audio;
     return 1;
 }